Add function in script.js file
