+++ package.json
@@ -12,8 +12,8 @@
     "style": "npx stylelint \"**/*.{css,scss}\"",
     "stylef": "npx stylelint --fix \"**/*.{css,scss}\"",
     "watch": "webpack --watch",
-    "start": "webpack-dev-server --config webpackConfig.js ",
-    "build": "webpack --config webpackConfig.js"
+    "start": "webpack-dev-server --config webpack.config.js",
+    "build": "webpack --config webpack.config.js"
   },
   "repository": {
     "type": "git",
@@ -31,22 +31,22 @@
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.25.4",
+    "file-loader": "^6.2.0",
+    "hint": "^6.1.9",
+    "html-loader": "^3.1.0",
+    "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
-    "stylelint": "^14.2.0",
-    "stylelint-config-standard": "^24.0.0",
-    "stylelint-csstree-validator": "^2.0.0",
-    "stylelint-scss": "^4.1.0",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.5.4",
-    "webpack": "^5.65.0",
+    "stylelint": "^13.13.1",
+    "stylelint-config-standard": "^21.0.0",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "^3.21.0",
+    "svg-url-loader": "^7.1.1",
+    "webpack": "^5.66.0",
     "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.7.2"
+    "webpack-dev-server": "^4.7.3"
   },
   "dependencies": {
-    "express": "^4.17.2",
-    "html-webpack-plugin": "^5.5.0",
-    "lodash": "^4.17.21",
-    "webpack-dev-middleware": "^5.3.0",
-    "xml-loader": "^1.2.1"
-  }
+    "lodash": "^4.17.21"
+  },
+  "main": "webpack.config.js"
 }
